+++ src/main/java/org/team4/hanzip/global/security/jwt/JwtProvider.java
@@ -8,7 +8,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 import org.team4.hanzip.domain.member.entity.Member;
-import org.team4.hanzip.global.security.CustomUserDetails;
 
 import javax.crypto.SecretKey;
 import java.util.Date;
@@ -60,19 +59,6 @@ public class JwtProvider {
                 .setExpiration(expiration)
                 .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
-    }
-
-    public Authentication getAuthentication(String accessToken) {
-        Jws<Claims> claims = Jwts
-                .parser()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(accessToken);
-        Long memberId = parseLong(claims.getBody().getSubject());
-
-        CustomUserDetails userDetails = new CustomUserDetails(memberId);
-        return new UsernamePasswordAuthenticationToken(memberId, null, null);
-
     }
 
 }
